+++ src/algorithms/pollards_rho.rs
@@ -21,27 +21,27 @@ impl Factorize for PollardsRho {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utils;
+    use crate::test_framework::prime_factorization::CheckTestBuilder;
 
     #[test]
     fn composites() {
-        test_utils::prime_factorization::CheckTestBuilder::<PollardsRho>::new()
+        CheckTestBuilder::<PollardsRho>::new()
             .case(8051, &[83, 97])
             .case(15, &[3, 5])
             .case(4096, &[2; 12])
             .build()
-            .check()
+            .check_cases()
     }
 
     #[test]
     fn primes() {
-        test_utils::prime_factorization::CheckTestBuilder::<PollardsRho>::new()
+        CheckTestBuilder::<PollardsRho>::new()
             .case(3, &[3])
             .case(5, &[5])
             .case(19, &[19])
             .case(29, &[29])
             .case(37, &[37])
             .build()
-            .check()
+            .check_cases()
     }
 }
