+++ src/test_framework.rs
@@ -1,0 +1,1 @@
+pub mod prime_factorization;
