+++ src/test_framework/prime_factorization.rs
@@ -1,0 +1,4 @@
+mod builder;
+mod check_test;
+
+pub use self::builder::CheckTestBuilder;
